Do not fail on Z_BUF_ERROR
The zlib documentation seems to hint that Z_BUF_ERROR is not
necessarily a result of incorrect parameters being passed to zlib, and
can in fact occur in normal use.

* zlib.h:

    Note that Z_BUF_ERROR is not fatal, and deflate() can be called
    again with more input and more output space to continue
    compressing.

* http://www.zlib.net/zlib_faq.html#faq05
* http://zlib.net/zlib_how.html

This change was motivated by a crash on a "buffer error"
TDTZCompressorException in one of our apps.  The crash is not
reproducible, and neither do we have the exact data or state of CocoaZ
at the time the crash occurred.  The caller code that uses CocoaZ has
been reasonably well tested, and does not have any "obvious" incorrect
parameter passing issues.